#2639: reverted MCTS files to r14016 to be able to merge this ticket
git-svn-id: https://src.heuristiclab.com/svn/core/trunk@14135 2abd9481-f8db-48e9-bd25-06bc13291c1b

+++ sources/HeuristicLab.Algorithms.DataAnalysis/3.4/MctsSymbolicRegression/ExpressionEvaluator.cs
@@ -511,7 +511,7 @@ namespace HeuristicLab.Algorithms.DataAnalysis.MctsSymbolicRegression {
       op = code[pc++];
       s = 0;
       if (op == (byte)OpCodes.LoadVar) {
-        s = (short)((code[pc] << 8) | code[pc + 1]);
+        s = (short)(((short)code[pc] << 8) | (short)code[pc + 1]);
         pc += 2;
       }
     }
